# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Space Reservation API",
        "description": "API REST para sistema de reserva de espacios para eventos. Permite a los usuarios explorar espacios disponibles, hacer reservas y gestionarlas.",
        "contact": {
            "name": "Soporte API",
            "email": "admin@espacios.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/spaces": {
            "get": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Listar todos los espacios",
                "operationId": "4d73a69d5f22f345d4f45ba91c45f052",
                "parameters": [
                    {
                        "name": "capacity_min",
                        "in": "query",
                        "description": "Capacidad mínima",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "capacity_max",
                        "in": "query",
                        "description": "Capacidad máxima",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar por estado activo",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre o ubicación",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de espacios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Space"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Crear nuevo espacio (Solo Admin)",
                "operationId": "26f76a00e23873c125497f7651535f8a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "capacity",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nueva Sala de Juntas"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Sala moderna con equipamiento"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Edificio D - Piso 1"
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "WiFi",
                                            "Proyector"
                                        ]
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/sala.jpg"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Espacio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Espacio creado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Space"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (solo admin)"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/spaces/{id}": {
            "get": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Obtener detalle de un espacio",
                "operationId": "51263b3f4e4a7c14833461ffcba86833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Space"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Actualizar espacio (Solo Admin)",
                "operationId": "a2d349d3652a2f3082fc09d61880814e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sala Actualizada"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "image_url": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Espacio actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Espacio actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Space"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Eliminar espacio (Solo Admin)",
                "operationId": "6b10994fdab655b7e3083432631bab0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espacio eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Espacio eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/spaces/{id}/availability": {
            "get": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Verificar disponibilidad de un espacio",
                "operationId": "14637e7343c2aeb5cb33e06b405b09cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Fecha y hora de inicio (Y-m-d H:i:s)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "Fecha y hora de fin (Y-m-d H:i:s)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de disponibilidad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "available": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "space_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_time": {
                                            "type": "string",
                                            "example": "2025-01-15 09:00:00"
                                        },
                                        "end_time": {
                                            "type": "string",
                                            "example": "2025-01-15 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Space": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sala de Conferencias Principal"
                    },
                    "description": {
                        "type": "string",
                        "example": "Sala amplia con capacidad para 30 personas"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 30
                    },
                    "location": {
                        "type": "string",
                        "example": "Edificio A - Piso 2"
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "WiFi",
                            "Proyector",
                            "Pizarra",
                            "Aire Acondicionado"
                        ]
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/sala.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token de autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}